# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger for the API that was developed as a core platform of the BUILDSPACE project.\n**Note** that the **X-Group-Id** is a header that is required in all the enpoints.",
        "title": "BUILSPACE Core Platform Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "BUILDSPACE Core Platform Support",
            "url": "http://www.swagger.io/support",
            "email": "isotiropoulos@singularlogic.eu"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/bucket": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Use a Bucket model to create a new bucket.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buckets"
                ],
                "summary": "Create bucket.",
                "parameters": [
                    {
                        "description": "Bucket payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Bucket"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bucket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            }
        },
        "/bucket/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a bucket based on it's ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buckets"
                ],
                "summary": "Delete bucket with all contents.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bucket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Bucket": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.ErrorReport": {
            "type": "object",
            "properties": {
                "internal_status": {
                    "description": "Status of the error",
                    "type": "string"
                },
                "message": {
                    "description": "Message of the error",
                    "type": "string"
                },
                "reason": {
                    "description": "Reason of the error",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the error",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}